+++ iotdm-application/src/main/java/com/aibaixun/iotdm/controller/ProductController.java
@@ -7,7 +7,9 @@ import com.aibaixun.basic.result.JsonResult;
 import com.aibaixun.iotdm.entity.Product;
 import com.aibaixun.iotdm.service.IProductService;
 import com.aibaixun.iotdm.support.ProductInfo;
+import com.aibaixun.iotdm.util.UserInfoUtil;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -68,7 +70,14 @@ public class ProductController extends BaseController{
     }
 
     @DeleteMapping("/{id}")
-    public JsonResult<Boolean> removeProduct (@PathVariable String id) {
+    public JsonResult<Boolean> removeProduct (@PathVariable String id) throws BaseException {
+        Product pr = productService.getById(id);
+        if (Objects.isNull(pr)){
+            throw new BaseException("产品不存在无法删除",BaseResultCode.GENERAL_ERROR);
+        }
+        if (!StringUtils.equals(pr.getCreator(), UserInfoUtil.getUserIdOfNull())){
+            throw new BaseException("产品必须由创建人删除",BaseResultCode.GENERAL_ERROR);
+        }
         boolean remove = productService.removeById(id);
         return JsonResult.success(remove);
     }
